+++ tests/test_workspaces.rs
@@ -216,8 +216,8 @@ fn test_list_workspaces_template() {
     test_env.jj_cmd_success(test_env.env_root(), &["init", "--git", "main"]);
     test_env.add_config(
         r#"
-        template.commit_summary = """commit_id.short() " " description.first_line()
-                                     if(current_working_copy, " (current)")"""
+        templates.commit_summary = """commit_id.short() " " description.first_line()
+                                      if(current_working_copy, " (current)")"""
         "#,
     );
     let main_path = test_env.env_root().join("main");